pyproject.toml
[project]
name = "last-fm-mcp"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "httpx>=0.28.1",
    "mcp[cli]>=1.13.1",
    "pydantic>=2.11.7",
    "python-dotenv>=1.1.1",
]
